report the condition a refused size belongs to
A size past the 4-byte SSZ offset is ErrSszSizeExceeded: no target encodes
it. A size only this target's int cannot hold is ErrPlatformOverflow, which
a target with a wider int would have held. sszutils.SizeLimitSentinel decides
between them from the size, and every size-domain report on both engines
routes through it, through ErrPlatformOverflowFn, or through
ErrSszSizeLimitFn, which applies the same rule to a count and an element
width by dividing the limit instead of forming the product.

ErrPlatformOverflowFn takes a uint64 so it can weigh its value the same way.

A delegate's refusal is passed on by a generated sizer instead of summed
into a plausible total.

A refusal carrying no figure states no size, so it cannot claim a wider int
would have held one, and reports the size limit on every target. Where the
platform's int is the narrower bound, that leaves a generated path that
pre-sizes naming the size limit where reflection still separates the two.

+++ reflection/treeroot.go
@@ -655,7 +655,7 @@ func (ctx *ReflectionCtx) buildRootFromUnion(sourceType *ssztypes.TypeDescriptor
 func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescriptor, sourceValue reflect.Value, hh sszutils.HashWalker, depth reflectionDepth) error {
 	vecLen := sourceType.Len
 	if vecLen > math.MaxInt {
-		return sszutils.ErrPlatformOverflowFn("vector length", sourceType.Len)
+		return sszutils.ErrPlatformOverflowFn("vector length", uint64(sourceType.Len))
 	}
 
 	packed := packedElemSize(sourceType.ElemDesc) > 0
